refactor(platform): redistribute config tests to domain modules (#2830)
Move each single-domain test cluster into its module (audio/video/debugger/
cartridge) and the CLI-framework tests into cli, sharing the test_support
helpers. mod.rs keeps only cross-cutting and orchestrator tests. No behavior
change; full platform suite still green (339 tests).
