+++ src/platform/config/mod.rs
@@ -330,133 +330,11 @@ pub(crate) mod test_support {
 mod tests {
     use super::test_support::{config_new, parse_config};
     use super::*;
-    use crate::nes::console::{Config, HardwareModel};
+    use crate::nes::console::HardwareModel;
 
-    // Help tests
-    #[test]
-    fn test_config_help_flag() {
-        let args = vec!["neser".to_string(), "--help".to_string()];
-        match config_new(args).unwrap() {
-            ParseResult::Help => {}
-            ParseResult::Version => panic!("Expected Help, got Version"),
-            ParseResult::Config(_) => panic!("Expected Help"),
-        }
-    }
-
-    #[test]
-    fn test_config_help_flag_short() {
-        let args = vec!["neser".to_string(), "-h".to_string()];
-        match config_new(args).unwrap() {
-            ParseResult::Help => {}
-            ParseResult::Version => panic!("Expected Help, got Version"),
-            ParseResult::Config(_) => panic!("Expected Help"),
-        }
-    }
-
-    #[test]
-    fn test_config_version_flag_returns_version_before_validation() {
-        let args = vec![
-            "neser".to_string(),
-            "--version".to_string(),
-            "--not-a-real-flag".to_string(),
-        ];
-
-        match Config::new(&args).unwrap() {
-            ParseResult::Version => {}
-            ParseResult::Help => panic!("Expected Version, got Help"),
-            ParseResult::Config(_) => panic!("Expected Version, got Config"),
-        }
-    }
-
-    #[test]
-    fn test_help_text_lists_version_flag() {
-        let help = cli::help_text();
-
-        assert!(help.contains("--version"));
-        assert!(help.contains("Print version information and exit"));
-    }
-
-    #[test]
-    fn test_help_text_groups_flags_into_readable_sections() {
-        let help = cli::help_text();
-
-        assert!(help.contains("\nInput:"));
-        assert!(help.contains("\nTrace and Debugging:"));
-        assert!(help.contains("\nSound:"));
-        assert!(help.contains("\nVideo and Display:"));
-        assert!(help.contains("\nAutorun:"));
-        assert!(help.contains("\nCartridge Catalog:"));
-
-        let input_section = help.find("\nInput:").unwrap();
-        let input_flag = help.find("--nes-controller-port1").unwrap();
-        assert!(input_section < input_flag);
-
-        let trace_section = help.find("\nTrace and Debugging:").unwrap();
-        let trace_flag = help.find("--trace-cpu").unwrap();
-        assert!(trace_section < trace_flag);
-
-        let sound_section = help.find("\nSound:").unwrap();
-        let sound_flag = help.find("--audio").unwrap();
-        assert!(sound_section < sound_flag);
-    }
-
-    #[test]
-    fn test_help_text_load_state_is_presence_only_flag() {
-        let help = cli::help_text();
-
-        assert!(help.contains("--load-state"));
-        assert!(!help.contains("--no-load-state"));
-        assert!(!help.contains("--disable-load-state"));
-    }
-
-    #[test]
-    fn test_help_text_lists_audio_buffer_ms_flag() {
-        let help = cli::help_text();
-
-        assert!(help.contains("--audio-buffer-ms"));
-        assert!(help.contains("Target native audio buffering in milliseconds"));
-    }
-
-    #[test]
-    fn test_help_text_lists_audio_sample_rate_flag() {
-        let help = cli::help_text();
-
-        assert!(help.contains("--audio-sample-rate"));
-        assert!(help.contains("Target native audio sample rate in Hz"));
-        assert!(help.contains("22050, 44100, 48000, 96000, 192000"));
-        assert!(help.contains("default: 44100"));
-    }
-
-    #[test]
-    fn test_help_text_oam_dram_decay_shows_default_and_no_negation_aliases() {
-        let help = cli::help_text();
-
-        assert!(help.contains("--nes-oam-dram-decay"));
-        assert!(help.contains("default: false"));
-        assert!(!help.contains("--no-oam-dram-decay"));
-        assert!(!help.contains("--disable-oam-dram-decay"));
-    }
-
-    #[test]
-    fn test_help_text_examples_use_hardware_flag() {
-        let help = cli::help_text();
-
-        assert!(help.contains("neser --nes-hardware nes-pal game.nes"));
-        assert!(!help.contains("--tv-system"));
-    }
-
-    // Audio/Video/Input CLI tests
-    #[test]
-    fn test_config_vsync_false() {
-        let args = vec![
-            "neser".to_string(),
-            "--vsync".to_string(),
-            "false".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(!config.frontend.vsync_enabled);
-    }
-
+    // Cross-cutting and orchestrator tests. Single-domain tests live in the
+    // corresponding domain modules (audio, video, autorun, debugger, cartridge)
+    // and the CLI framework tests live in `cli`.
     #[test]
     fn test_config_gamepads_false() {
         let args = vec![
@@ -466,86 +344,6 @@ mod tests {
         ];
         let config = parse_config(args);
         assert!(!config.frontend.gamepads_enabled);
-    }
-
-    #[test]
-    fn test_config_fullscreen_true() {
-        let args = vec![
-            "neser".to_string(),
-            "--fullscreen".to_string(),
-            "true".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(config.frontend.fullscreen);
-        assert_eq!(config.frontend.fullscreen_display, None);
-    }
-
-    #[test]
-    fn test_config_load_state_flag() {
-        let args = vec!["neser".to_string(), "--load-state".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.load_state);
-    }
-
-    #[test]
-    fn test_config_fullscreen_with_display() {
-        let args = vec![
-            "neser".to_string(),
-            "--fullscreen".to_string(),
-            "true".to_string(),
-            "--display".to_string(),
-            "1".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(config.frontend.fullscreen);
-        assert_eq!(config.frontend.fullscreen_display, Some(1));
-    }
-
-    #[test]
-    fn test_config_display_without_fullscreen_is_ignored() {
-        let args = vec![
-            "neser".to_string(),
-            "--display".to_string(),
-            "1".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(!config.frontend.fullscreen);
-        assert_eq!(config.frontend.fullscreen_display, None);
-    }
-
-    #[test]
-    fn test_config_display_missing_value_errors() {
-        let args = vec![
-            "neser".to_string(),
-            "--fullscreen".to_string(),
-            "--display".to_string(),
-        ];
-        let result = config_new(args);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_config_display_invalid_value_errors() {
-        let args = vec![
-            "neser".to_string(),
-            "--fullscreen".to_string(),
-            "--display".to_string(),
-            "abc".to_string(),
-        ];
-        let result = config_new(args);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_config_display_negative_value_errors() {
-        let args = vec![
-            "neser".to_string(),
-            "--fullscreen".to_string(),
-            "--display".to_string(),
-            "-1".to_string(),
-        ];
-        let result = config_new(args);
-        assert!(result.is_err());
     }
 
     #[test]
@@ -561,147 +359,6 @@ mod tests {
             result.unwrap_err(),
             "Invalid filter name: 'invalid-filter'. Valid options are: none, crt, smooth, ntsc, pal"
         );
-    }
-
-    #[test]
-    fn test_config_debugger_enabled() {
-        let args = vec![
-            "neser".to_string(),
-            "--debugger".to_string(),
-            "true".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(config.frontend.debugger_enabled);
-    }
-
-    #[test]
-    fn test_config_unknown_argument_errors() {
-        let args = vec![
-            "neser".to_string(),
-            "--definitely-not-a-real-flag".to_string(),
-        ];
-        let result = config_new(args);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_config_positional_argument_is_rom_path() {
-        let args = vec!["neser".to_string(), "somefile.nes".to_string()];
-        let config = parse_config(args);
-        assert_eq!(config.frontend.rom_path.as_deref(), Some("somefile.nes"));
-    }
-
-    // Tracing CLI tests
-    #[test]
-    fn test_config_tracing_enabled() {
-        let args = vec!["neser".to_string(), "--trace".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.cpu, 1); // --trace enables CPU tracing at level 1
-    }
-
-    #[test]
-    fn test_config_tracing_nestest() {
-        let args = vec!["neser".to_string(), "--trace-nestest".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert!(config.frontend.tracing.nestest);
-    }
-
-    #[test]
-    fn test_config_tracing_cpu() {
-        let args = vec!["neser".to_string(), "--trace-cpu".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.cpu, 1);
-    }
-
-    #[test]
-    fn test_config_tracing_ppu() {
-        let args = vec!["neser".to_string(), "--trace-ppu".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.ppu, 1);
-    }
-
-    #[test]
-    fn test_config_tracing_apu() {
-        let args = vec!["neser".to_string(), "--trace-apu".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.apu, 1);
-    }
-
-    #[test]
-    fn test_config_tracing_mapper() {
-        let args = vec!["neser".to_string(), "--trace-mapper".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.mapper, 1);
-    }
-
-    #[test]
-    fn test_config_tracing_cpu_with_level() {
-        let args = vec!["neser".to_string(), "--trace-cpu=2".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.cpu, 2);
-    }
-
-    #[test]
-    fn test_config_tracing_ppu_with_level() {
-        let args = vec!["neser".to_string(), "--trace-ppu=3".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.ppu, 3);
-    }
-
-    #[test]
-    fn test_config_tracing_ppu_level_is_capped_at_five() {
-        let args = vec!["neser".to_string(), "--trace-ppu=9".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.ppu, 5);
-    }
-
-    #[test]
-    fn test_config_tracing_apu_with_level() {
-        let args = vec!["neser".to_string(), "--trace-apu=4".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.apu, 4);
-    }
-
-    #[test]
-    fn test_config_tracing_mapper_with_level() {
-        let args = vec!["neser".to_string(), "--trace-mapper=5".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.mapper, 5);
-    }
-
-    #[test]
-    fn test_config_tracing_mapper_level_is_capped_at_five() {
-        let args = vec!["neser".to_string(), "--trace-mapper=9".to_string()];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.mapper, 5);
-    }
-
-    #[test]
-    fn test_config_tracing_with_multiple_levels() {
-        let args = vec![
-            "neser".to_string(),
-            "--trace-cpu=3".to_string(),
-            "--trace-ppu=2".to_string(),
-            "--trace-apu=1".to_string(),
-        ];
-        let config = parse_config(args);
-        assert!(config.frontend.tracing.enabled);
-        assert_eq!(config.frontend.tracing.cpu, 3);
-        assert_eq!(config.frontend.tracing.ppu, 2);
-        assert_eq!(config.frontend.tracing.apu, 1);
-        assert_eq!(config.frontend.tracing.mapper, 0);
     }
 
     #[test]
@@ -732,182 +389,9 @@ mod tests {
         assert!(!config.nes.apu_channels.contains(ApuChannels::NOISE));
     }
 
-    #[test]
-    fn test_config_window_height() {
-        let args = vec![
-            "neser".to_string(),
-            "--window-height".to_string(),
-            "720".to_string(),
-        ];
-        let config = parse_config(args);
-        assert_eq!(config.frontend.window_height, 720);
-    }
-
-    #[test]
-    fn test_config_window_height_invalid_errors() {
-        let args = vec![
-            "neser".to_string(),
-            "--window-height".to_string(),
-            "not_a_number".to_string(),
-        ];
-        let result = config_new(args);
-        assert!(result.is_err());
-    }
-
-    // Config file tests
-    // Note: Config file tests that use apply_config_value remain in nes::console::config
-    // since apply_config_value is private and config file parsing is orchestrated there.
-    // These CLI tests verify the platform flags work correctly.
-
-    #[test]
-    fn test_config_metadata_db_path_defaults_to_none() {
-        let config = parse_config(vec!["neser".to_string(), "game.nes".to_string()]);
-        assert!(config.frontend.metadata_db_path.is_none());
-    }
-
-    #[test]
-    fn test_config_metadata_db_path_from_cli() {
-        let config = parse_config(vec![
-            "neser".to_string(),
-            "--metadata-db-path".to_string(),
-            "/custom/metadata.db".to_string(),
-            "game.nes".to_string(),
-        ]);
-        assert_eq!(
-            config.frontend.metadata_db_path.as_deref(),
-            Some("/custom/metadata.db")
-        );
-    }
-
-    #[test]
-    fn test_config_image_cache_path_defaults_to_none() {
-        let config = parse_config(vec!["neser".to_string(), "game.nes".to_string()]);
-        assert!(config.frontend.image_cache_path.is_none());
-    }
-
-    #[test]
-    fn test_config_image_cache_path_from_cli() {
-        let config = parse_config(vec![
-            "neser".to_string(),
-            "--image-cache-path".to_string(),
-            "/custom/cache".to_string(),
-            "game.nes".to_string(),
-        ]);
-        assert_eq!(
-            config.frontend.image_cache_path.as_deref(),
-            Some("/custom/cache")
-        );
-    }
-
-    #[test]
-    fn test_config_metadata_db_path_from_config_file() {
-        use std::io::Write;
-        use tempfile::NamedTempFile;
-
-        let mut file = NamedTempFile::new().unwrap();
-        file.write_all(b"metadata_db_path=/from/config/metadata.db\n")
-            .unwrap();
-
-        let config = parse_config(vec![
-            "neser".to_string(),
-            "--config".to_string(),
-            file.path().to_string_lossy().to_string(),
-            "game.nes".to_string(),
-        ]);
-        assert_eq!(
-            config.frontend.metadata_db_path.as_deref(),
-            Some("/from/config/metadata.db")
-        );
-    }
-
-    #[test]
-    fn test_config_image_cache_path_from_config_file() {
-        use std::io::Write;
-        use tempfile::NamedTempFile;
-
-        let mut file = NamedTempFile::new().unwrap();
-        file.write_all(b"image_cache_path=/from/config/cache\n")
-            .unwrap();
-
-        let config = parse_config(vec![
-            "neser".to_string(),
-            "--config".to_string(),
-            file.path().to_string_lossy().to_string(),
-            "game.nes".to_string(),
-        ]);
-        assert_eq!(
-            config.frontend.image_cache_path.as_deref(),
-            Some("/from/config/cache")
-        );
-    }
-
-    #[test]
-    fn test_config_cli_overrides_config_file_metadata_db_path() {
-        use std::io::Write;
-        use tempfile::NamedTempFile;
-
-        let mut file = NamedTempFile::new().unwrap();
-        file.write_all(b"metadata_db_path=/from/config/metadata.db\n")
-            .unwrap();
-
-        let config = parse_config(vec![
-            "neser".to_string(),
-            "--config".to_string(),
-            file.path().to_string_lossy().to_string(),
-            "--metadata-db-path".to_string(),
-            "/from/cli/metadata.db".to_string(),
-            "game.nes".to_string(),
-        ]);
-        assert_eq!(
-            config.frontend.metadata_db_path.as_deref(),
-            Some("/from/cli/metadata.db")
-        );
-    }
-
-    #[test]
-    fn test_resolved_metadata_db_path_uses_configured_value() {
-        let cfg = FrontendConfig {
-            metadata_db_path: Some("/custom/metadata.db".to_string()),
-            ..Default::default()
-        };
-        assert_eq!(
-            cfg.resolved_metadata_db_path(),
-            std::path::PathBuf::from("/custom/metadata.db")
-        );
-    }
-
-    #[test]
-    fn test_resolved_metadata_db_path_falls_back_to_default() {
-        let cfg = FrontendConfig::default();
-        let path = cfg.resolved_metadata_db_path();
-        assert!(
-            path.ends_with(".neser/metadata.db"),
-            "expected path ending with .neser/metadata.db, got: {path:?}"
-        );
-    }
-
-    #[test]
-    fn test_resolved_image_cache_path_uses_configured_value() {
-        let cfg = FrontendConfig {
-            image_cache_path: Some("/custom/cache".to_string()),
-            ..Default::default()
-        };
-        assert_eq!(
-            cfg.resolved_image_cache_path(),
-            std::path::PathBuf::from("/custom/cache")
-        );
-    }
-
-    #[test]
-    fn test_resolved_image_cache_path_falls_back_to_default() {
-        let cfg = FrontendConfig::default();
-        let path = cfg.resolved_image_cache_path();
-        assert!(
-            path.ends_with(".neser/image_cache"),
-            "expected path ending with .neser/image_cache, got: {path:?}"
-        );
-    }
-
+    // Config-file value parsing that exercises `apply_config_value` directly is
+    // covered here and in `nes::console::config`; the orchestrator normalizes
+    // dashed keys to underscores before delegating to the domain modules.
     #[test]
     fn test_apply_config_value_accepts_dashes_for_underscore_keys() {
         let mut cfg = FrontendConfig::default();
